+++ sydtest-mutation-plugin/src/Test/Syd/Mutation/Plugin/Operators.hs
@@ -25,6 +25,7 @@ import qualified Test.Syd.Mutation.Plugin.Operator.MaybeOp
 import qualified Test.Syd.Mutation.Plugin.Operator.Negate
 import qualified Test.Syd.Mutation.Plugin.Operator.RemoveAction
 import qualified Test.Syd.Mutation.Plugin.Operator.RemoveCase
+import qualified Test.Syd.Mutation.Plugin.Operator.SwitchFunctionArguments
 import Test.Syd.Mutation.Plugin.Operators.TH (collectOperators)
 
 allOperators :: [MutationOperator]